day4/agents/requirements.txt
pydantic>=2.0.0
asyncio>=3.4.3
streamlit>=1.32.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
python-dotenv>=1.0.0 